put logic in DateTimeHelper
