+++ app/src/main/java/com/example/tasky/agenda_details/presentation/utils/DateTimeHelper.kt
@@ -5,7 +5,9 @@ import java.time.Instant
 import java.time.LocalDate
 import java.time.LocalTime
 import java.time.ZoneId
+import java.time.ZoneOffset
 import java.time.ZonedDateTime
+import java.time.format.DateTimeFormatter
 
 object DateTimeHelper {
     fun getLocalDateFromEpoch(
@@ -32,5 +34,13 @@ object DateTimeHelper {
 
         val duration = Duration.between(startTimeInSeconds, endTimeInSeconds)
         return duration.seconds
+    }
+
+    fun getEpochSecondsFromDateAndTime(date: String, time: String): Long {
+        val dateFormatter = DateTimeFormatter.ofPattern("MMM dd yyyy")
+        val localTime = LocalTime.parse(time)
+        val localDate = LocalDate.parse(date, dateFormatter)
+        val zdt = ZonedDateTime.of(localDate, localTime, ZoneOffset.UTC)
+        return zdt.toEpochSecond()
     }
 }
